Refactor: refactored remaining code
