+++ packages/tangle/messagefactory.go
@@ -67,7 +67,7 @@ func (f *MessageFactory) SetWorker(worker Worker) {
 // IssuePayload creates a new message including sequence number and tip selection and returns it.
 // It also triggers the MessageConstructed event once it's done, which is for example used by the plugins to listen for
 // messages that shall be attached to the tangle.
-func (f *MessageFactory) IssuePayload(p payload.Payload, t ...*OldTangle) (*Message, error) {
+func (f *MessageFactory) IssuePayload(p payload.Payload, t ...*Tangle) (*Message, error) {
 	payloadLen := len(p.Bytes())
 	if payloadLen > payload.MaxSize {
 		err := fmt.Errorf("maximum payload size of %d bytes exceeded", payloadLen)
@@ -94,7 +94,7 @@ func (f *MessageFactory) IssuePayload(p payload.Payload, t ...*OldTangle) (*Mess
 	// due to the ParentAge check we must ensure that we set the right issuing time.
 	if t != nil {
 		for _, parent := range strongParents {
-			t[0].Message(parent).Consume(func(msg *Message) {
+			t[0].Storage.Message(parent).Consume(func(msg *Message) {
 				if msg.ID() != EmptyMessageID && !msg.IssuingTime().Before(issuingTime) {
 					time.Sleep(msg.IssuingTime().Sub(issuingTime) + 1*time.Nanosecond)
 					issuingTime = clock.SyncedTime()
