+++ plugins/issuer/plugin.go
@@ -32,7 +32,7 @@ func configure(_ *node.Plugin) {}
 
 // IssuePayload issues a payload to the message layer.
 // If the node is not synchronized an error is returned.
-func IssuePayload(payload payload.Payload, t ...*tangle.OldTangle) (*tangle.Message, error) {
+func IssuePayload(payload payload.Payload, t ...*tangle.Tangle) (*tangle.Message, error) {
 	if !syncbeaconfollower.Synced() {
 		return nil, xerrors.Errorf("can't issue payload: %w", syncbeaconfollower.ErrNodeNotSynchronized)
 	}
